+++ packages/firebase-dart/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 7.0.0-dev
 
 - **BREAKING** renamed `AuthCredential` into `OAuthCredential` to align with JS API
+- **BREAKING** removed deprecated Firestore `Settings.timestampsInSnapshots`. 
 - Added `idToken`, `accessToken` and `secret` to `OAuthCredential`
 - Added support for Email Link authentication.
 - Firestore `Settings` added `cacheSizeBytes`, `host`, and `ssl` properties.